-add: font settings
Added ability to select your own fonts for PacketList, FieldGrid View and RawData View from the program settings.
Changed the RawDataView to be better at auto-resizing to enable the previews more easy.
Changes to preview, so it now shows the value for what is has found a match.
Updated version