-add: popup for rules editor, local var view
Added popup menu to the rule editor to allow for easy editing.
It now also moves and selects the text as it seems fit for easier editing.
When you now right-click on the field view, you can select between showing parsed Fields, or the resulting local variable stack of the parsed packet (if it has a rule attached).
You can also pick "Show debug information" to show the Field view including debug information of actions that don't normally produce a output line.