-add: change rules of loaded project
Added ability to change the active rules for a open project.
When opening a file without a project, or if no valid reader could be found, the program will now prompt what rules you want to use (it there is more than one option)
Fixed a bug where loading a new rule would still maintain some of the old rule data if the packet rule was not present in the newly loaded rule.