-add: vpx export
Added copy to clipboard, and export to .vpx file functions
VPX stands for Viewed Packet Export. This is a bzip2 compressed xml file with a 4-byte header in front of it.
The export will only save the packets that are currently visible in the list. (filtered packets are not saved)