-fix: decompression, main menu
Added LibVLC to the installer script to make sure you can play a video.
Changed ``Search`` in the main menu to ``Search and Filters``
The menus should no longer be disabled when you open a file, after closing the welcome tab.
Fixed a bug related to decompression where I forgot it initialize the decompressor.