-fix: PacketList Scrollbar bug
There is a bug with ListBox in that moving the scrollbar by dragging with a mouse is broken if you have more than 64K items in the list.
I added a workaround for this by faking a scrollbar that is linked to the actual ListBox. To do this, I also had to further customize the ListBox by adding a ShowScrollBar property that allows to fully hide and disable it's vertical scrollbar.