-update: Copy Summary
Added a checkbox to Copy Summary in the project settings to only show the names or IDs of the packet that are currently visible instead of all of them. This can be used to remove irrelevant packets from the summary.
Also fixed a bug where Decryption Key-Name wasn't saved correctly into the project settings file.